Move get_comparison_baseline to azure_util
